fix: detect `await Promise<Result>` correctly
When checking AwaitExpression, we were checking the type of `node.argument` (the Promise),
which is not Result-like for `Promise<Result>`.
We should check the type of the AwaitExpression itself (the awaited value),
which will be `Result` and thus correctly detected as Result-like.

Also removed recursion in `isHandledResult` for AwaitExpression
as it's better to check if the AwaitExpression node itself is handled by its parent.